requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==37.0.2
h11==0.13.0
idna==3.2
munch==2.5.0
munch-stubs==0.1.1
mypy==0.950
mypy-extensions==0.4.3
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
requests==2.26.0
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
types-pytz==2021.3.8
types-PyYAML==6.0.7
types-requests==2.27.25
types-urllib3==1.26.14
typing_extensions==4.2.0
urllib3==1.26.6
wsproto==1.1.0
